feat: ClassMapGenerator::scanPaths can exclude specific dirs (#7)
* feat: ClassMapGenerator::scanPaths can exclude specific dirs

Part of https://github.com/composer/class-map-generator/issues/6

Class Finder from Symfony is used to get an iterator on files matching
the first argument of scanPaths. Currently if we want to exclude some
files from the search we can provide a regex which will be test on each
file.

Finder has a method exclude() which allow us to exclude a list of dirs
relative to the one passed in in(). It does a pre-filtering much more
efficient than checking the regex.

To keep backward compatibility, this new argument is added at then end
of the method with a default value.