Simplify `PhpFileParser::getExtraTypes()` logic
The first time the method is called, a string is built up with `|`-separated
extra types, then split based on `|`s, the resulting array is filtered to
exclude non-empty strings, and that array is finally passed to
`PhpFileCleaner::setTypeConfig()`. Given that the array only has at most a
single entry, just set the array directly.